Route splitting/api call
