+++ views/layouts/main.handlebars
@@ -84,6 +84,10 @@
         {{/if}}
         {{!-- End flash messages --}}
 
+        <div class="alert alert-success d-none" role="alert" id="alert-success"></div>
+        <div class="alert alert-danger d-none" role="alert" id="alert-error"></div>
+        <div class="alert alert-info d-none" role="alert" id="alert-info"></div>
+
         {{{body}}}
     </main>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
