+++ views/mylibrary.handlebars
@@ -1,6 +1,4 @@
-<h1>This is the My Library page</h1><br><br>
-
-  <div class="pb-3">
+<div class="pb-3">
     <div>Have a Book to Swap? <a href="/addbook">Add Book Now</a></div>
   </div>
 <div class="shelf">
